Add: BuildingSystem namespace Refactored: Building folder structure Renamed: BuildingComponent to Component Renamed: BuildingComposite to Composite Renamed: BuildingFactory to ComponentFactory Renamed: BuidlingSystem to ThreadsafeFacade Moved: Composite Pattern related classes to folder Building/CompositePattern
