+++ src/Building/HttpAdapter.cpp
@@ -1,242 +1,241 @@
-#pragma clang diagnostic push
-#pragma ide diagnostic ignored "cert-err34-c"
+
 //
 // Created by Julia on 2019-12-30.
 //
 
 
-#include <http-server/Building/HttpAdapter.h>
+#include "http-server/Building/HttpAdapter.h"
 
 using namespace boost::xpressive;
 
-
-HttpAdapter::HttpAdapter(BuildingSystem &buildingSystem) : buildingSystem(buildingSystem)
+namespace BuildingSystem
 {
-    get_regex = sregex::compile(
-            "^GET /buildings"
-            "("
-            /**/"/(?P<building>[[:digit:]]+)"
-            /**/"("
-            /*  */"/floors/(?P<floor>[[:digit:]]+)"
-            /*  */"(/rooms/(?P<room>[[:digit:]]+))?"
-            /**/")?"
-            ")?"
-            "/(?P<what>equipment|structure)"
-    );
-    put_regex = sregex::compile(
-            "^PUT /buildings"
-            "("
-            /**/"/(?P<building>[[:digit:]]+)/floors"
-            /**/"("
-            /*  */"/(?P<floor>[[:digit:]]+)/rooms"
-            /*      */"("
-            /*      */"/(?P<room>[[:digit:]]+)/equipment"
-            /*  */")?"
-            /**/")?"
-            ")?$"
-    );
-
-    post_regex = sregex::compile(
-            "^POST /buildings(?P<building>[[:digit:]]+)"
-            "/floors/(?P<floor>[[:digit:]]+)"
-            "/rooms/(?P<room>[[:digit:]]+)"
-            "/equipment/(?P<equipment>[[:digit:]]+)"
-            "$"
-    );
-    destination_regex = sregex::compile(
-            "^/buildings(?P<building>[[:digit:]]+)"
-            "/floors/(?P<floor>[[:digit:]]+)"
-            "/rooms/(?P<room>[[:digit:]]+)"
-            "$"
-    );
-    delete_regex = sregex::compile(
-            "^DELETE /buildings/(?P<building>[[:digit:]]+)"
-            "("
-            /**/"/floors/(?P<floor>[[:digit:]]+)"
-            /**/"("
-            /*  */"/rooms/(?P<room>[[:digit:]]+)"
-            /*  */"("
-            /*      */"/equipment/(?P<equipment>[[:digit:]]+)"
-            /*  */")?"
-            /**/")?"
-            ")?$"
-    );
-}
-
-void HttpAdapter::operator()(int connection_socket_descriptor)
-{
-    HttpRequestReader *httpRequestReader = new BsdSocket_HttpRequestReader(connection_socket_descriptor);
-    HttpResponseBuilder httpResponseBuilder;
-    request = httpRequestReader->get_request();
-
-    std::cout << "@\t\tRECIEVED A REQUEST FROM: " << connection_socket_descriptor << "\n"
-              << request.get_request_line() << "\n";
-
-    HttpResponse response = respond_to_request();
-
-    std::string strResponse = response.to_str();
-
-    std::cout << "@\t\tSENDING A RESPONSE TO: " << connection_socket_descriptor << "\n"
-              << response.get_response_line() << "\n";
-
-    write(connection_socket_descriptor, strResponse.c_str(), strResponse.size());
-
-    delete httpRequestReader;
-}
-
-HttpResponse HttpAdapter::respond_to_request()
-{
-    responseBuilder.init();
-
-    std::string str = request.get_request();
-    smatch match_path;
-    if (regex_search(str, match_path, get_regex))
+    HttpAdapter::HttpAdapter(ThreadsafeFacade &buildingSystem) : buildingSystem(buildingSystem)
     {
-        respond_to_get(match_path);
+        get_regex = sregex::compile(
+                "^GET /buildings"
+                "("
+                /**/"/(?P<building>[[:digit:]]+)"
+                /**/"("
+                /*  */"/floors/(?P<floor>[[:digit:]]+)"
+                /*  */"(/rooms/(?P<room>[[:digit:]]+))?"
+                /**/")?"
+                ")?"
+                "/(?P<what>equipment|structure)"
+        );
+        put_regex = sregex::compile(
+                "^PUT /buildings"
+                "("
+                /**/"/(?P<building>[[:digit:]]+)/floors"
+                /**/"("
+                /*  */"/(?P<floor>[[:digit:]]+)/rooms"
+                /*      */"("
+                /*      */"/(?P<room>[[:digit:]]+)/equipment"
+                /*  */")?"
+                /**/")?"
+                ")?$"
+        );
+
+        post_regex = sregex::compile(
+                "^POST /buildings(?P<building>[[:digit:]]+)"
+                "/floors/(?P<floor>[[:digit:]]+)"
+                "/rooms/(?P<room>[[:digit:]]+)"
+                "/equipment/(?P<equipment>[[:digit:]]+)"
+                "$"
+        );
+        destination_regex = sregex::compile(
+                "^/buildings(?P<building>[[:digit:]]+)"
+                "/floors/(?P<floor>[[:digit:]]+)"
+                "/rooms/(?P<room>[[:digit:]]+)"
+                "$"
+        );
+        delete_regex = sregex::compile(
+                "^DELETE /buildings/(?P<building>[[:digit:]]+)"
+                "("
+                /**/"/floors/(?P<floor>[[:digit:]]+)"
+                /**/"("
+                /*  */"/rooms/(?P<room>[[:digit:]]+)"
+                /*  */"("
+                /*      */"/equipment/(?P<equipment>[[:digit:]]+)"
+                /*  */")?"
+                /**/")?"
+                ")?$"
+        );
     }
-    else if (regex_search(str, match_path, post_regex))
-    {
-        respond_to_post(match_path);
-    }
-    else if (regex_search(str, match_path, delete_regex))
-    {
-        respond_to_delete(match_path);
-    }
-    else if (regex_search(str, match_path, put_regex))
-    {
-        respond_to_put(match_path);
 
-    }
-    else
+    void HttpAdapter::operator()(int connection_socket_descriptor)
     {
-        responseBuilder.set_body("Request not recognized\n");
-        responseBuilder.set_status_code(StatusCode::Not_Implemented);
+        HttpRequestReader *httpRequestReader = new BsdSocket_HttpRequestReader(connection_socket_descriptor);
+        HttpResponseBuilder httpResponseBuilder;
+        request = httpRequestReader->get_request();
+
+        std::cout << "@\t\tRECIEVED A REQUEST FROM: " << connection_socket_descriptor << "\n"
+                  << request.get_request_line() << "\n";
+
+        HttpResponse response = respond_to_request();
+
+        std::string strResponse = response.to_str();
+
+        std::cout << "@\t\tSENDING A RESPONSE TO: " << connection_socket_descriptor << "\n"
+                  << response.get_response_line() << "\n";
+
+        write(connection_socket_descriptor, strResponse.c_str(), strResponse.size());
+
+        delete httpRequestReader;
     }
-    responseBuilder.setHeaderInfo({{"Server",       {"BuildingSystem"}},
-                                   {"Connection",   {"Keep-Alive"}},
-                                   {"Content-Type", {"text/xml"}}
-                                  });
-    return responseBuilder.getResponse();
-}
 
-void HttpAdapter::respond_to_put(const smatch &match_path)
-{
-    int source_building = to_int(match_path, "building");
-    int source_floor = to_int(match_path, "floor");
-    int source_room = to_int(match_path, "room");
-
-    try
+    HttpResponse HttpAdapter::respond_to_request()
     {
-        json j = json::parse(request.get_body());
+        responseBuilder.init();
 
-        if (source_room != 0)
+        std::string str = request.get_request();
+        smatch match_path;
+        if (regex_search(str, match_path, get_regex))
         {
-            std::shared_ptr<Equipment> equipment = j;
-            buildingSystem.add({source_building, source_floor, source_room}, equipment);
+            respond_to_get(match_path);
+        }
+        else if (regex_search(str, match_path, post_regex))
+        {
+            respond_to_post(match_path);
+        }
+        else if (regex_search(str, match_path, delete_regex))
+        {
+            respond_to_delete(match_path);
+        }
+        else if (regex_search(str, match_path, put_regex))
+        {
+            respond_to_put(match_path);
+
         }
         else
         {
-            std::shared_ptr<BuildingComponent> buildingComponent = j;
-            buildingSystem.add({source_building, source_floor, source_room}, buildingComponent);
+            responseBuilder.set_body("Request not recognized\n");
+            responseBuilder.set_status_code(StatusCode::Not_Implemented);
         }
-        responseBuilder.set_body("Item created\n");
-        responseBuilder.set_status_code(StatusCode::OK);
+        responseBuilder.setHeaderInfo({{"Server",       {"BuildingSystem"}},
+                                       {"Connection",   {"Keep-Alive"}},
+                                       {"Content-Type", {"text/xml"}}
+                                      });
+        return responseBuilder.getResponse();
+    }
 
-    }
-    catch (...)
+    void HttpAdapter::respond_to_put(const smatch &match_path)
     {
-        responseBuilder.set_body("Something went wrong\n");
-        responseBuilder.set_status_code(StatusCode::Bad_Request);
-    }
-}
+        int source_building = to_int(match_path, "building");
+        int source_floor = to_int(match_path, "floor");
+        int source_room = to_int(match_path, "room");
 
-void HttpAdapter::respond_to_delete(const smatch &match_path)
-{
-    int source_building = to_int(match_path, "building");
-    int source_floor = to_int(match_path, "floor");
-    int source_room = to_int(match_path, "room");
-    int source_equipment = to_int(match_path, "equipment");
-    try
-    {
-        buildingSystem.remove({source_building, source_floor, source_room, source_equipment});
-        responseBuilder.set_body("Item deleted.\n");
-        responseBuilder.set_status_code(StatusCode::OK);
-    }
-    catch (...)
-    {
-        responseBuilder.set_body("Something went wrong\n");
-        responseBuilder.set_status_code(StatusCode::Bad_Request);
-    }
-    /*CO NAJMNIEJ BUILDING PODANE*/}
-
-void
-HttpAdapter::respond_to_post(const smatch &match_path)
-{
-    auto values = request.get_field_value("destination");
-    if (values.size() == 1 && values[0] != HttpMessage::NO_SUCH_KEY)
-    {
-        smatch match_destination;
-        if (regex_search(values[0], match_destination, destination_regex))
+        try
         {
-            int source_building = to_int(match_path, "building");
-            int source_floor = to_int(match_path, "floor");
-            int source_room = to_int(match_path, "room");
-            int source_equipment = to_int(match_path, "equipment");
+            json j = json::parse(request.get_body());
 
-
-            int destination_building = to_int(match_path, "building");
-            int destination_floor = to_int(match_path, "floor");
-            int destination_room = to_int(match_path, "room");
-
-            try
+            if (source_room != 0)
             {
-                buildingSystem.move(source_equipment, {source_building, source_floor, source_room},
-                                    {destination_building, destination_floor, destination_room});
-                responseBuilder.set_status_code(StatusCode::OK);
+                std::shared_ptr<Equipment> equipment = j;
+                buildingSystem.add({source_building, source_floor, source_room}, equipment);
             }
-            catch (...)
+            else
             {
-                responseBuilder.set_body("Equipment couldn't be moved\n");
-                responseBuilder.set_status_code(StatusCode::Not_Modified);
+                std::shared_ptr<Component> buildingComponent = j;
+                buildingSystem.add({source_building, source_floor, source_room}, buildingComponent);
+            }
+            responseBuilder.set_body("Item created\n");
+            responseBuilder.set_status_code(StatusCode::OK);
+
+        }
+        catch (...)
+        {
+            responseBuilder.set_body("Something went wrong\n");
+            responseBuilder.set_status_code(StatusCode::Bad_Request);
+        }
+    }
+
+    void HttpAdapter::respond_to_delete(const smatch &match_path)
+    {
+        int source_building = to_int(match_path, "building");
+        int source_floor = to_int(match_path, "floor");
+        int source_room = to_int(match_path, "room");
+        int source_equipment = to_int(match_path, "equipment");
+        try
+        {
+            buildingSystem.remove({source_building, source_floor, source_room, source_equipment});
+            responseBuilder.set_body("Item deleted.\n");
+            responseBuilder.set_status_code(StatusCode::OK);
+        }
+        catch (...)
+        {
+            responseBuilder.set_body("Something went wrong\n");
+            responseBuilder.set_status_code(StatusCode::Bad_Request);
+        }
+        /*CO NAJMNIEJ BUILDING PODANE*/}
+
+    void
+    HttpAdapter::respond_to_post(const smatch &match_path)
+    {
+        auto values = request.get_field_value("destination");
+        if (values.size() == 1 && values[0] != HttpMessage::NO_SUCH_KEY)
+        {
+            smatch match_destination;
+            if (regex_search(values[0], match_destination, destination_regex))
+            {
+                int source_building = to_int(match_path, "building");
+                int source_floor = to_int(match_path, "floor");
+                int source_room = to_int(match_path, "room");
+                int source_equipment = to_int(match_path, "equipment");
+
+
+                int destination_building = to_int(match_path, "building");
+                int destination_floor = to_int(match_path, "floor");
+                int destination_room = to_int(match_path, "room");
+
+                try
+                {
+                    buildingSystem.move(source_equipment, {source_building, source_floor, source_room},
+                                        {destination_building, destination_floor, destination_room});
+                    responseBuilder.set_status_code(StatusCode::OK);
+                }
+                catch (...)
+                {
+                    responseBuilder.set_body("Equipment couldn't be moved\n");
+                    responseBuilder.set_status_code(StatusCode::Not_Modified);
+                }
+            }
+            else
+            {
+                responseBuilder.set_body("Inproper destination path.\n");
+                responseBuilder.set_status_code(StatusCode::Not_Found);
             }
         }
         else
         {
-            responseBuilder.set_body("Inproper destination path.\n");
-            responseBuilder.set_status_code(StatusCode::Not_Found);
+            responseBuilder.set_body("Expected a destination header, got none.\n");
+            responseBuilder.set_status_code(StatusCode::Precondition_Failed);
         }
     }
-    else
+
+    void HttpAdapter::respond_to_get(const smatch &match_path)
     {
-        responseBuilder.set_body("Expected a destination header, got none.\n");
-        responseBuilder.set_status_code(StatusCode::Precondition_Failed);
+        int building = to_int(match_path, "building");
+        int floor = to_int(match_path, "floor");
+        int room = to_int(match_path, "room");
+        std::__cxx11::string what = match_path["what"].str();
+
+        try
+        {
+            std::string body = (what == "equipment" ? buildingSystem.get_equipment({building, floor, room}) :
+                                buildingSystem.get_structure({building, floor, room}));
+            responseBuilder.set_body(body);
+            responseBuilder.set_status_code(StatusCode::OK);
+        }
+        catch (...)
+        {
+            responseBuilder.set_body("Something went wrong\n");
+            responseBuilder.set_status_code(StatusCode::Bad_Request);
+        }
+    }
+
+    int HttpAdapter::to_int(const smatch &match_path, const std::string &name) const
+    {
+        return atoi(match_path[name].str().c_str());
     }
 }
-
-void HttpAdapter::respond_to_get(const smatch &match_path)
-{
-    int building = to_int(match_path, "building");
-    int floor = to_int(match_path, "floor");
-    int room = to_int(match_path, "room");
-    std::__cxx11::string what = match_path["what"].str();
-
-    try
-    {
-        std::string body = (what == "equipment" ? buildingSystem.get_equipment({building, floor, room}) :
-                            buildingSystem.get_structure({building, floor, room}));
-        responseBuilder.set_body(body);
-        responseBuilder.set_status_code(StatusCode::OK);
-    }
-    catch (...)
-    {
-        responseBuilder.set_body("Something went wrong\n");
-        responseBuilder.set_status_code(StatusCode::Bad_Request);
-    }
-}
-
-int HttpAdapter::to_int(const smatch &match_path, const std::string &name) const
-{
-    return atoi(match_path[name].str().c_str());
-}
-
-#pragma clang diagnostic pop
